+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - add nested diagnostic locations in node CLI (#51)
 - fix lexer crash on certain ASCII control characters (#294)
 - fix duplicate output for `--no-color` option
+- fix inconsistencies between native and node CLI output (#307)
 - fix performance and memory leak issues related to `regex`
 - fix incorrect handling of final sigma characters in `cowel_to_lower` (#169)
 - fix various code block syntax highlighting issues by updating µlight